[WIP] Add an extension for OpenID-based authentication.
This introduces OpenID-based authentication to Review Board 2.0. It
extends the login page to provide options for logging in using common
services, such as GitHub or Google. It accepts any OpenID provider.

This is still a work-in-progress. It needs the in-development
Review Board 2.0 beta 2 changes, and doesn't yet provide anything for
the My Account page.

This is heavily based on the work done by Patrick Uiterwijk.